refactor: move storage layout constants
