+++ orchestrator/src/core/client/storage/mod.rs
@@ -1,8 +1,5 @@
-pub mod constants;
 pub mod error;
 pub mod s3;
-
-pub use constants::*;
 
 use async_trait::async_trait;
 use bytes::Bytes;
